+++ packages/state-manager/src/sagas/reducers.ts
@@ -10,6 +10,7 @@ import { settingsReducer } from './settings/settings.slice'
 import { StoreKeys } from './store.keys'
 import { usersReducer } from './users/users.slice'
 import { captchaReducer } from './captcha/captcha.slice'
+import { reactionsReducer } from './reactions/reactions.slice'
 
 export const reducers = {
   [StoreKeys.PublicChannels]: publicChannelsReducer,
@@ -23,4 +24,5 @@ export const reducers = {
   [StoreKeys.Files]: filesReducer,
   [StoreKeys.Network]: networkReducer,
   [StoreKeys.Captcha]: captchaReducer,
+  [StoreKeys.Reactions]: reactionsReducer,
 }